requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
aiofiles
PyMuPDF
pdfplumber
pymongo
motor
camelot-py[cv]
ollama
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-readers-file
llama-index-vector-stores-chroma
chromadb
paddlepaddle
paddleocr
pillow
numpy
